+++ src/main/java/com/jajaja/global/config/security/jwt/JwtProperties.java
@@ -1,4 +1,4 @@
-package com.jajaja.global.config.security;
+package com.jajaja.global.config.security.jwt;
 
 import lombok.Getter;
 import lombok.Setter;
